effort chart of COCOMO estimation

+++ src/estimation/detailedEstimate/COCOMOEstimate.java
@@ -40,7 +40,7 @@ public class COCOMOEstimate extends ParameterArea{
 		createEffortMultipliers(form.getBody());
 	}
 	
-	public double getEstimatedSize()
+	public int getEstimatedSize()
 	{
 		return sizeSpinner.getSelection();
 	}
